<refactor>: Separate validation methods
car 클래스에 위치했던 자동차 이름에 대한 검증 메서드를 carnamevalidator 클래스로 분리
